Reordered text in the "Receiving Responses" section.
I think the text flows better if we talk about the TO=0 case
first and the TO=1 case second so that we can refer handling a TLS
handshake failure back to the TO=0 case.

I removed John's erroneous addition that the client could send a
query over TCP after a TLS handshake failure.