Addressing comments from Stephane Bortzmeyer:
I did not find a serious problem. I have one question and one
criticism.

> Since pipelined responses can arrive out-of-order, clients MUST
> match responses to outstanding queries using the ID field and port
> number.

I do not understand how this works. All replies on a given TCP
connection will have the same source port (the new well-known port)
and the same destination port (the one used to open the TCP
connection). So, how do you use the port number for demultiplexing?
Why not using the QNAME instead? (The query ID may be unsufficient if
there are a lot of outstanding queries + the birthday paradox.)

> For DNS clients that use library functions such as
> "gethostbyname()",

This was replaced by a better function in RFC 2133, in 1997...