Farrell:
- 4.1, RFC 7435 is about "opportunistic security" and not about
"opportunistic encryption." That followed a long terminology debate,
so you might want to s/encryption/security/ there. (But I don't care
personally:-)

- 4.1, last para: OS never provides "guaranteed" privacy. I think
what you mean is that while you're always vulnerable to an on-path
attacker, if there is no such attacker then you do get privacy (to
the recursive) albeit with no guarantee. I think that's subtly
different from what's stated.

- 4.2, "The user MUST be alerted" needs a qualification - what if
it's my calendar running in the background? I'd prepend a "Where
possible,... " at least.

- Section 7: I'm not clear from the note if you do or don't want
this text in the RFC. I'd suggest keeping it. (Except the note.)

- Section 9, point 4: I think it'd be fair to note that naive
padding schemes may not work as well as planned against traffic
analysis, and to say that implementations that allow for flexibility
there might be good so that we can experiment with ways of
mitigating traffic analysis.

- Section 10: This contradicts the front page which has 6 names
listed.